python-pretender/requirements.txt
avatar2
PyYAML
pyserial
statsmodels
numpy
scipy
scikit-learn
